requirements.txt
anyio==4.1.0
certifi==2023.11.17
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
openpyxl==3.1.2
python-decouple==3.8
python-telegram-bot==20.7
sniffio==1.3.0
Unidecode==1.3.7
